fix fifo usage calculation

+++ upf/upf_proxy.c
@@ -489,7 +489,7 @@ common_fifo_tuning_callback (session_t * s, svm_fifo_t * f,
   u8 seg_usage = fifo_segment_get_mem_usage (fs);
   u32 fifo_in_use = svm_fifo_max_dequeue_prod (f);
   u32 fifo_size = svm_fifo_size (f);
-  u8 fifo_usage = fifo_in_use * 100 / fifo_size;
+  u32 fifo_usage = ((u64)fifo_in_use * 100) / (u64)fifo_size;
   u32 update_size = 0;
 
   ASSERT (act < SESSION_FT_ACTION_N_ACTIONS);
